+++ Cinema/src/main/java/com/example/Cinema/service/ProgrammeService.java
@@ -1,6 +1,8 @@
 package com.example.Cinema.service;
 
 import com.example.Cinema.exception.MovieNotFoundException;
+import com.example.Cinema.exception.ProgrammeNotFoundException;
+import com.example.Cinema.mapper.ProgrammeMapper;
 import com.example.Cinema.model.CinemaHall;
 import com.example.Cinema.model.dto.ProgrammeDto;
 import com.example.Cinema.model.Movie;
@@ -8,6 +10,7 @@ import com.example.Cinema.model.Programme;
 import com.example.Cinema.repository.CinemaHallRepository;
 import com.example.Cinema.repository.MovieRepository;
 import com.example.Cinema.repository.ProgrammeRepository;
+import com.example.Cinema.service.Validators.ProgrammeValidationService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -20,12 +23,16 @@ public class ProgrammeService {
     private final ProgrammeRepository programmeRepository;
     private final MovieRepository movieRepository;
     private final CinemaHallRepository cinemaHallRepository;
+    private final ProgrammeMapper programmeMapper;
+    private final ProgrammeValidationService  programmeValidationService;
 
     @Autowired
-    public ProgrammeService(ProgrammeRepository programmeRepository, MovieRepository movieRepository, CinemaHallRepository cinemaHallRepository) {
+    public ProgrammeService(ProgrammeRepository programmeRepository, MovieRepository movieRepository, CinemaHallRepository cinemaHallRepository, ProgrammeMapper programmeMapper, ProgrammeValidationService programmeValidationService) {
         this.programmeRepository = programmeRepository;
         this.movieRepository = movieRepository;
         this.cinemaHallRepository = cinemaHallRepository;
+        this.programmeMapper = programmeMapper;
+        this.programmeValidationService = programmeValidationService;
     }
 
     public void save(Programme programme) {
@@ -33,11 +40,18 @@ public class ProgrammeService {
     }
 
     public void deleteById(Long id) {
+        programmeValidationService.validateEditPermission(id);
+
         programmeRepository.deleteById(id);
     }
 
-    public Optional<Programme> getProgrammeById(Long id) {
-        return programmeRepository.findById(id);
+    public Programme getProgrammeById(Long id) {
+        return programmeRepository.findById(id).orElseThrow(() -> new ProgrammeNotFoundException(id));
+    }
+
+    public ProgrammeDto getProgrammeDtoById(Long id) {
+        Programme programme = programmeRepository.findById(id).orElseThrow(() -> new ProgrammeNotFoundException(id));
+        return programmeMapper.toDto(programme);
     }
 
     public List<Programme> getProgrammes(LocalDate date) {
@@ -48,7 +62,9 @@ public class ProgrammeService {
         return programmeRepository.findByTitleDateHallName(title, date, hallName);
     }
 
-    public void updateProgramme(ProgrammeDto programmeDto) {
+    public void update(ProgrammeDto programmeDto) {
+        programmeValidationService.validateEditPermission(programmeDto.getIdprogramme());
+        programmeValidationService.validateCinemaHallAvailable(programmeDto);
 
         Programme programme = programmeRepository.findById(programmeDto.getIdprogramme()).orElse(new Programme());
